+++ NEWS.md
@@ -1,5 +1,7 @@
 # 0.4.1 (dev)
-- improved vertical line drawing for time ranges 10 to 20 years (one line every year)
+- added a new argument `showLabels` to choose whether or not the event labels shall be drawn - improves layout of dense timelines
+- new argument `lineInterval`: the distance in seconds that vertical lines shall be drawn (to reduce plot size and increase performance). When omitted, a heuristic is used (as before) is used.
+- improved heuristic of vertical line drawing for time ranges 10 to 20 years (one line every year)
 - corrected font colors of Presidents example on help page
 
 # 0.4.0 (CRAN)